koh/witchquiz/witchquiz/witchquiz/signals.py
from django.conf import settings
from django.db.models.signals import post_save
from django.dispatch import receiver
from rest_framework.authtoken.models import Token
from witchquiz.models import UserToken

@receiver(post_save, sender=settings.AUTH_USER_MODEL)
def create_auth_token(sender, instance=None, created=False, **kwargs):
    if created:
        token = UserToken.objects.get(user=instance)
        token = Token.objects.create(user=instance, key=token.token)