koh/witchquiz/witchquiz/pyproject.toml
[tool.poetry]
name = "witchquiz"
version = "0.1.0"
description = ""
authors = ["theoremoon <theoldmoon0602@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
apscheduler = "^3.10.0"
numpy = "^1.24.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
